requirements.txt
pytesseract
pdf2image
python-docx
pillow
langchain
langchain-community
pypdf
google-generativeai
fastapi
uvicorn
